Clarify uiTableModelRowInserted API documentation.
* Require the new row to have been added to the model
  before calling uiTableModelRowInserted.
* Require the row count to be updated before calling
  uiTableModelRowInserted.

Notes
-----

darwin: does not care about row counts internally. From the docs:
        > The numberOfRows in the table view is automatically increased
        > by the count of indexes.

unix: does not state anything regarding order in the documentation.

windows: does not state anything in the documentation but does keep
         track of its own row count internally.

So in theory we do not have to require increasing the row count before
calling uiTableModelRowInserted, but just by the very nature of the
function name (past tense) we should.